requirements.txt
sklearn-crfsuite
nltk